Refactor renderInBanner, Do board refresh on GameScreen (more consistant that way)
